Add files via upload
This project creates a basic calculator which checks the user inputs. Making sure they're integers the program will return friendlier error messages if they are not. It also checks the denominator if the user chooses division, making sure the input is not zero, if it is the program will return a friendly error message.